refactor: update package structure for user registration DTOs and handlers
