+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/user/register/dto/mapper/UserMapperDto.java
@@ -1,7 +1,6 @@
-package co.com.crediya.api.user.dto.mapper;
+package co.com.crediya.api.user.register.dto.mapper;
 
-import co.com.crediya.api.user.dto.request.CreateUserRequestDto;
-import co.com.crediya.api.user.dto.response.CreateUserResponseDto;
+import co.com.crediya.api.user.register.dto.response.CreateUserResponseDto;
 import co.com.crediya.model.user.User;
 import org.mapstruct.Mapper;
