+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/user/register/dto/response/CreateUserResponseDto.java
@@ -1,4 +1,4 @@
-package co.com.crediya.api.user.dto.response;
+package co.com.crediya.api.user.register.dto.response;
 
 public record CreateUserResponseDto(
         String name,
